Refactor AudioAPI util functionality into a React component.
